Delete obsolete test, merge into new torch workflow test
